backend/requirements.txt
django

djangorestframework

celery[redis]

psycopg2-binary
